+++ R/data_dictionary.R
@@ -42,6 +42,7 @@ data_dictionary <- function(year,
   # dataset = 'population'
 
   ### check inputs
+  if (missing(year) || is.null(year)) { error_year_not_declared() }
   checkmate::assert_numeric(year, any.missing = FALSE)
   checkmate::assert_string(dataset, na.ok = FALSE)
   checkmate::assert_logical(verbose, null.ok = FALSE)