+++ R/read_families.R
@@ -32,6 +32,7 @@ read_families <- function(year,
                           verbose = TRUE){
 
   ### check inputs
+  if (missing(year) || is.null(year)) { error_year_not_declared() }
   checkmate::assert_numeric(year, any.missing = FALSE)
   checkmate::assert_vector(columns, null.ok = TRUE)
   checkmate::assert_logical(as_data_frame, null.ok = FALSE)